tests/robotests/src/com/android/car/settings/CarSettingActivityTest.java
/*
 * Copyright (C) 2016 The Android Open Source Project
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.android.car.settings;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.robolectric.annotation.Config;

import static com.google.common.truth.Truth.assertThat;

@RunWith(CarSettingsRobolectricTestRunner.class)
@Config(manifest = TestConfig.MANIFEST_PATH, sdk = TestConfig.SDK_VERSION)
public class CarSettingActivityTest {

    private CarSettingActivity mActivity;

    @Test
    public void testConstant() {
        mActivity = new CarSettingActivity();
        assertThat(mActivity.getSettingAction()).isEqualTo("android.car.settings.SETTINGS");
        assertThat(mActivity.getSettingPkg()).isEqualTo("com.android.car.settings");
    }
}
